Updated stuff that UUID is better than name
If the commit name doesn't make sense it's normal im claqué and braindead

There's bug wit tabManager but momo asked me to play valorant so i haven't fixed it

+++ src/mctdl/game/utils/TimeFormater.java
@@ -2,13 +2,13 @@ package mctdl.game.utils;
 
 public class TimeFormater {
 	
-public static String getFormatedTime(long left) {
+public static String getFormatedTime(int left) {
 		
 		String format = "";
 		
-		long dividend = left, divisor = 60;
-        long quotient = dividend / divisor;
-        long remainder = dividend % divisor;
+		int dividend = left, divisor = 60;
+        int quotient = dividend / divisor;
+        int remainder = dividend % divisor;
         
         String hour, min, secs;
